Migrated project to NextJS
