+++ components/MemorialComponents/SkyArea.js
@@ -1,4 +1,3 @@
-import React from 'react'
 import styled from 'styled-components'
 
 const SkyAreaWrapper = styled.div`
