+++ components/ResourcesComponents/HomeListing.js
@@ -1,4 +1,3 @@
-import React from 'react'
 import styled from 'styled-components'
 
 const HomeListingWrapper = styled.div`
