--- src/components/MemorialComponents/MemorialContext.js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-const MemorialContext = React.createContext();
-
-export default MemorialContext;
