--- src/components/ResourcesComponents/ResourcesContext.js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-const ResourcesContext = React.createContext();
-
-export default ResourcesContext;
